Move layout selector into the extensions directory
