+++ src/extensions/layout-selector/editor.scss
