+++ src/styles/editor.scss
@@ -6,7 +6,6 @@
 @import "../components/**/editor.scss";
 @import "../extensions/**/editor.scss";
 @import "../blocks/**/editor.scss";
-@import "../layout-selector/style.scss";
 
 // Legacy styling
 @import "legacy/margin-editor";
